module set up for future export.  TS configured for build.  Sending up to merge with GUI folder.
