+++ NPM-MODULE/tsconfig.json
@@ -16,6 +16,5 @@
         ]
       }
     },
-    "include": ["src"],
     "exclude": ["node_modules", "**/__tests__/*"]
   }
